Return API_EINTERNAL at command::procresult if chatroom not found
When we proccess a request before send it, we have sanity checks to
check if chatroom exists or not (in this case we use to return
API_ENOENT).

If we have sent (a chat related) command we can assume that chatroom
exists at that moment, so upon command::procresult, it still should
exists, otherwise we can consider as API_EINTERNAL as chatroom must
exists at that point.