Remove SIC for the restoration of invalid/incomplete fingerprints
Node attributes are not updated in Node objects before they have been
applied on the server side.

Due to these changes, if there is a transient error causing repeated
calls to Transfer::complete, fingerprints could be restored on each
call. To prevent this, apart from the current fingerprint of the node,
now pending attribute changes for the node are also checked.

Check if the fingerprint is already being fixed before trying to fix it
again