Fix race condition at mCompletionForMegaApiThread (folder controllers)
At some point of folder controllers algorithm execution, we are
performing some operations on 'worker' thread, and we need a thread
change, to continue executing algorithm on SDK's thread.

To accomplish this requirement, we call executeOnThread, and we push a
new request type on requestQueue, when this request is processed, we'll
execute lambda function provided as param in executeOnThread.

The problem is when this request is processed after we have logged out,
as so many resources has been freed/removed.

To fix this problem, we will capture a weak_ptr in
mCompletionForMegaApiThread lambdas, so in case
MegaFolderUpload/DownloadController are already destroyed we will just
return.

Note that all transfers will be destroyed when we logged, out upon
logout_result, when abortPendingActions is called.