SDK-3180. Notify app of errors for new Syncs/Backups
For newly created syncs and backups, the notifyApp flag is initially set
to false. This flag serves to notify the app of new errors or state
changes related to the sync/backup.

There exist situations where a sync/backup is created with an error that
cannot be identified prior to its creation. For instance, an error is
triggered when the root path for a sync is read-only, preventing the
creation of the megaignore file. In such cases, and any other where
errors are triggered upon creation, the notifyApp flag needs to be set
to true.

Given that this flag is solely utilized to communicate changes in the
states of syncs/backups, and considering the above rationale,
maintaining this flag as a parameter on MegaClient::addsync() is
unwarranted.

This commit removes the notifyApp parameter from MegaClient::addsync(),
and directly sets this flag to true when invoking
Syncs::appendNewSync().