SDK-3282. Externally modified backups should always start on mirroring mode when resumed
The actual behavior for a new or resumed backup is to start in mirroring
mode in the following cases:

1. When it doesn't have any state (i.e., it's the first time this
backup is running)
    2. When the backup is an external backup
    3. When it was disabled because it was externally modified.

However, point 3 has a bug: the value used for this check is mKnownError
(because mError is reset just before this check), and mKnownError is not
serialized, nor updated to the value of mError when this one is
unserialized.

Therefore, if a backup is externally modified and disabled because of
this, but the app has been restarted before it is resumed again, then
the backup would start in monitoring mode instead of mirroring mode,
breaking the expected behavior.

This commit fixes this scenario.