SDK-3225. Upgrade to SRW from Prod should discard any existing SRW cache
In SRW, the DB cache storing the state of the sync (ie.
syncNode/cloudNode/fsNode) is loaded upon start and used to resume
syncs.

In the migration from prod version to SRW, it's not expected to find
such DB cache. However if the user tried a SRW version in the past, then
there could be leftovers from previous runs, which will create issues.

This commit increases the DB version for SRW, and adapt the code
similarly to NOD:
1. If upgrading from NO SRW to SRW, the db files are just renamed to
the new DB version.
2. If downgrading from SRW to NO SRW, there will be files with the
increased DB version.
3. If upgrading again from NO SRW to SRW, those old files from
previous SRW version (which are now outdated) will be substituted by the
ones from NO SRW (which are updated).