Fix symlink syncs in SRW
1. The root sync path is now resolved by following symlinks.
Consequently, the fsid and other attributes correspond to the actual
path that the symlink references. This approach aligns with our
expectations to avoid path collisions, ensuring we retrieve the
statistics for the actual path, rather than the symlink itself.

2. In order to maintain pre-SRW logic (which doesn't follow symlinks
during scanning, meaning children of type symlink are not synced), the
scanDirectory() method has been adjusted. When the root path is a
symlink, the method will now resolve it, retrieving the statistics of
the actual path that the symlink references.