Remove support for schedid param in mcms command
With latest changes now API doesn't require schedid param to determine
if call must ring or not, and if waiting room must be enabled or not
for call.

Now a call will have waiting room enabled if chatroom flag is also
enabled

About ringing, call won't ring if notRinging param is true.

Finally note that:
- If call is ringing, all participants that answers the call, will
bypass the waiting room.

- If call has stopped ringing, or call was started with notRinging
param true, all participants will be redirected to waiting room,
as soon as they answer the call