Add functionality to get logged in user's Password Manager base node
The new function exposed MegaApi::getPasswordManagerBase entails the
creation of the Password Manager base node if it didn't exist.

The creation of a the Password Manager base is requested via `pwmp`
command and the new node is concurrency-safe created by the
backend.

Remarks:

- the new user attribute `pwmh` is modelled as a non-encrypted private
attribute but no preffix is expected by backend upon its request

- Password Manager Base node cannot be deleted, nor overwritten