Adjust restag for getattr of non-existing attrs
It is required in order to find the associated `MegaRequest` in the
intermediate layer. If the call to `getua()` sent the command to the
API, the `restag = tag` is done upon completion of the command, but in
this case the command is not sent, so it's needed to adjust the
`restag`.