SDK-3422. All syncs/backups are shown as "Issue found" when any of them is affected by a stall
The Backup Center is now updated on the new sync status through
heartbeats (using the 'sphb' command). However, the current
implementation treats the sync stalled state as a global condition for
all syncs, without differentiating between them.

Under the present system, when a stall issue arises, all active syncs
send a heartbeat with the status 'CommandBackupPutHeartBeat::STALLED'.
This causes the Backup Center to display 'Issues found' for all active
syncs, even if some of them are not actually stalled.

This commit addresses the issue by introducing specific information for
each stalled sync. Consequently, the heartbeat will only indicate the
'STALLED' status for a particular sync if it is indeed currently
stalled.