SDK-3441. Use SPState state TEMPORARY_DISABLED for suspended syncs
When a sync is not active anymore, "sp" command (CommandBackupPut) is
executed to inform about the new state. For paused syncs, the SPState
used was TEMPORARY_DISABLED. For suspended syncs, it is either FAILED or
DISABLED.

On the other hand, when a sync is requested to be paused by the apps,
the request state being used is SyncRunState::Suspend. Hence, we want
that a sync in suspended state is consistent with the legacy Pause
state. Besides, it is coherent that a "Suspended" state can be
considered as a TEMPORARY_DISABLED, different from "Disabled", which
means a more permanent state.

This commit does the following changes in BackupInfoSync::getSyncState:

1. A suspended sync (SyncRunState::Suspend) with no errors will also be
set as TEMPORARY_DISABLED, same as for SyncRunState::Pause.
2. Both SyncRunState::Suspend and SyncRunState::Disable state will only
return CommandBackupPut::FAILED if SyncError is GREATER than
NO_SYNC_ERROR, instead of being different. This is because when a sync
changes its state to suspended or disabled, the error is set to
UNLOADING_SYNC (as it will be unloaded), so the actual comparison is not
correct, as it is always returning CommandBackupPut::FAILED.