Fix error in PasswordManager integration test
The test always expected a newly created Password Node / Folder when
it could be that there was already an existing item with the same name
under that same path.

Test-wise, this could happen if the test fails at any point after the
creation of the Node and before its deletion. An alternative to
prevent the implemented approach would be to include the deletion in
the TearDown section / RAII of the test; however, this approach was
preferred since one of the goals of the test is to V&V the deletion