SDK-3449. Fix MoveOrRenameCannotOccur stall issues
- Changed ParentFolderDoesNotExist stall entries to point to the correct
target path, either cloud or local.
- DifferentFileOrFolderIsAlreadyPresent is now changed from the
MoveOrRenameCannotOccur stall entry type to a user intervention type
(file changed in both sides, cloud and locally)
- Combination of MoveOrRenameCannotOccur stall issues on
getSyncProblems() is now fixed.
- We always keep one of the PathProblems present when comparing Stall
entries (generally, there will be only one PathProblem in one of the
paths)
- It can happen that we have two different PathProblems for the same
path when comparing two combinable stall entries. This should only
happen if one of the problems is DestinationPathInUnresolvedArea. In
this case, we keep the other entry, which is the most meaningful one and
potentially solvable.