SDK-3505. Crash on getLocalPath()
Sync::checkCloudPathForMovesRenames can be called when row.syncNode is
nullptr.

While there are checks for row.syncNode's validity throughout the code,
one critical check is missing. This oversight leads to a call to
row.syncNode->getLocalPath(), resulting in undefined behavior and a
persistent crash in the app. The issue can only be resolved by
uninstalling MEGASync or stopping the sync from the Backup Center.

This commit adds the necessary check as well as some extra asserts.