SDK-3515. Increase waitonsyncs general timeout tolerance
Pending and queued actions (like putnodes) can last more than expected
when a several parallel executions and a lot of actions are called at
the same time.

While the specific timeout (which checks if the syncs are still active,
and is defined for each test, usually for several seconds) is fine, the
general timeout is too short for these situations. It can also happen
that the sync related actions start later than expected, so even if they
have a expected total duration, they are stopped by this general
timeout.

The actual timeout value is of 1 minute, and here it will be increased
to 5 minutes, which is more than enough time to take all these possible
scenarios into consideration, and it should be fine even with a lot more
new tests running in parallel.

Note that most executions will normally be finished much sooner due to
the specific timeout.