Add extra info to assert in SyncTest.MirroringInternalBackupResumesInMirroringMode
The following assert has failed, and it shouldn't:

ASSERT_TRUE(cb.client.syncs.syncConfigByBackupId(id, config))

We don't know at this point which is the exact value of the handle "id",
and we are not sure either whether the SyncVec has syncs or is empty.
This info is now added to the assert in case it fails again.