SDK-3456. Notification of Stall Issues
The sync-engine doesn't keep a record of Stalls. Instead, it constantly
clears the list of existing ones and generates a new one with the
currently detected stalls. As result, the app doesn't get reliable
notifications/callbacks to know if there are new stalls to show/attend
(in advance/smart mode), which forces the app to explicitly retrieve the
list of stalls periodically (or on demand, via a Refresh button).

After this commit, the conflict list and stall list size are calculated
when the syncs are within a conflict o stall state (and only during that
state). If the number increases o decreasing during the conflict/stall
state, fireOnGlobalSyncStateChanged() will be called to tell the apps
that they could refresh the stall list, so they can call
MegaApi::getMegaSyncStallList() to get the updated list.