Prevent asserting before restoring execution permissions in SyncTest.ChangingDirectoryPermissions
This test removes the folder permissions for execution, check some
conditions (a stall issue has been triggered, the stall is of the
correct type, etc) and then restores again the permissions for
execution. When one of those asserts fail before the permissions are
restored, the directory cannot be automatically removed during the
cleanup and we need to remove it manually. This commit fixes it by first
checking, then restoring, and then executing the assert termination if
needed.