SDK-3650. Enhance SpeedController accuracy
The SpeedController calculates both the speed and the average speed
using the following methods:

1. Speed Calculation: The speed is determined by utilizing a
circular buffer that gathers data over a set period (50 deciseconds).
However, this approach yields inaccurate results due to a flaw in its
implementation.
For example, if more than one decisecond has elapsed since the last
calculation (e.g., 2 deciseconds), the bytes corresponding to the
additional deciseconds are incorrectly added to the same buffer
position. In this scenario, the circular buffer loops through two
positions: the position corresponding to the current decisecond (ds 1)
and the one corresponding to the next decisecond (ds 2). However,
instead of properly accounting for the data accumulated over the 2
deciseconds (with half of the value for each position), the buffer
position corresponding to ds 1 is erroneously reset to zero, and the
data for the 2 deciseconds is stored in the position corresponding to ds
2. Consequently, when the buffer returns to the position previously set
to zero, it adds the new data for that decisecond, which will be added
to the total sum along with the data from the contiguous position in the
buffer (which was the one with a value corresponding to 2 ds), resulting
in an incorrect summation of data for 3 deciseconds instead of the
intended 2. This is just an example, but there are other inaccuracies.

2. Mean Speed Calculation: The mean speed computation also suffers
from inaccuracies, particularly in the initial seconds of measurement.
This occurs because the calculation does not include the initial
decisecond (decisecond 0) in the division process. As a result, for the
first two deciseconds, the total bytes for those deciseconds are divided
by 1 (since the difference between the current time and the initial time
is considered to be 1, instead of the correct value of 2), leading to an
inflated average speed calculation.

This code enhances the calculation for the circular mean speed and the
total mean speed. Besides, more meaningful variable names are used, as
well as detailed comments for several parts of the code.