SDK-3744. Investigate failure "some inshares were not removed" while cleaning up tests
SdkTestNodeAttributes create inshares for testing. When those inshares
have not been removed in time, the SdkTest::Cleanup method will try to
remove them, first by removing the contact, and then removing the node
if still present.

However, when the email is already inserted into the "alreadyRemoved"
collection, the loop in charge of the inshare removal continues to the
next iteration, without trying to remove the node. This isn't really
correct because, if the email is not tagged as "alreadyRemoved", the
cleanup code will try to remove it, and it should do it even if it is
already tagged as "alreadyRemoved", because it could have happened a few
lines above, within the loop in charge of removing the contacts.

Hence, even if the email is already tagged as "alreadyRemoved", the loop
shouldn't continue before trying to remove the node.