SDK-3759. Reduce noProgressCount logging frecuency
Sync loop is currently logging info when the no progress count is
increasing.

While it can be helpful to know if there has been any progress with
syncs, it is not really necessary to know every single update of the
no-progress count.

This commit fixes that log, so it only logs the info if there is a skip
for scanning, an early exit, or every 500 accumulated no-progress
counts.