Fix error parsing FetchAds command "adf" for invalid ads
When a user is allowed to fetch ads, and provides 1 or several
invalid/non-existent ads, the response will be an OK error code with
a JSON array containing -9/API_ENOENT (or its relevant error code) for
each of those requested ads.

Response results will be returned in the same order as in `adf` parameter.