SDK-3787. Avoid cloning a node with a zerokey
Context:
When putting a new node, the nodekey is obfuscated with the client's
master key, generating a new 32 bytes key which is sent to the API.

Issue:
The API will reject this key if the first 16 bytes are equal to the last
16 bytes, which means that the key is either a zero key or was generated
with a zero key (i.e., the transferkey used to produce the
filekey/nodekey was a zero key).

Notice that either if the nodekey is a zerokey (all bytes are zero) or
it was generated with a zerokey (in this case, its first 16 bytes will
be equal to the last 16 bytes), the obfuscated key which is sent to the
API will also suffer from this condition: the first 16 bytes will be
equal to the last 16 bytes.

How this issue is currently handled:
When this happens, we will receive an API_EKEY (-14) error for that
node, and the transfer will be restarted. When the transfer is
restarted, a new fresh transfer key will be generated, changing the
final filekey and the nodekey which will be used. This is, no matter
whether the zerokey was the transfer key, the file key or the node key,
all of them will be substituted after the new transfer.

What is not currently handled:
If the node was already cached and a new upload detects that node and
consider it as a valid node, this node will be cloned for the putnodes.
If the nodekey of that node was a zerokey (for example, if the
serialized data was corrupted or empty), this will generate an infinite
loop with no solution.

__________________

Solution:
This branch adds the following code to avoid the above:
1) Add logic to check whether a key is a zero key. This can be used
for 16-byte generator keys and for 32-byte file keys.
2) Check if a node to be cloned has a zero key: discard it if it is
the case.
3) Add asserts and logs to check the validity of the nodekey on
CommandPutNodes.