Use node error returned from API instead of API_ENOENT for transfer failure on CommandPutNodes
When putnodes fail, we receive an API response whose param "f" is empty,
and the transfer fails with API_ENOENT (not found).

However, when there is a specific node failure (for example, API_EKEY
-14 for a node using a zero key or a bad key), it would be more
approppriate to use the specific error for the transfer failure.

Otherwise, the error returned can be misleading:while it is true that
the put node cannot be found (it is empty because the operation didn't
work), that's a generic error which doesn't give information on the
specific problem leading to an empty response for putnodes.