SDK-3696. Crash on detecting duplicated nodes in backups
Sync::syncItem_checkBackupCloudNameClash moves duplicated backup cloud
items to cloud sync debris when duplicated items are still present after
previous operations.

When there is a syncNode, the "removeNodeHere" field is set to the
delete pointer, in order to prevent earlier resolution for delete
operations. However, when the syncNode doesn't exist and it is null,
accessing to the LocalNode::rare() method will result on a crash.

This commit fixes it by checking whether the syncNode exists before
accessing the LocalNode::rare() method.