Force upload if the existing node has a zerokey
For a regular upload, a previous node or a node with the same name is
compared to see if the actual upload could be avoided and we could just
clone the node.

That also need a check, and force the upload if the base node has a
zerokey. In that case, MegaApiImpl::hasToForceUpload should return true.

Also, an event should be send, similar to what we already do within
LocalNode::queueClientUpload for syncs.