Improve FileFingerprint serialization inline doc
There are a pair of very similar functions to serialize/deserialize
file fingerprints which are error prone to be mixed due to its similar
name, and exact same parameters. Additionally, they must be used in
pairs and the only way to identify which one to used is based on
context / checking which one was the first pair used.

While a long-lasting refactor arrives, this commit aims to help
developers with a minimal piece of wisdom/warning on this topic as
well as the content expected in each serialization.