Prevent a delete operation from one sync to become stalled due to pending operations from other syncs
This commit aims to provide a specific solution for the reported issue
where a delete operation is stalled due to pending move or rename
operations from other syncs, with minimum changes which don't affect the
current design (oriented to share states and operations across syncs).

In order to do this, delete operations (i.e., when a cloud node is gone,
or a FS node is gone, or the sync node must be deleted) will take into
account any pending moves only for the sync they belong.