Individualize scanningWasComplete flag for individual syncs
scanningWasComplete flag is also shared across syncs, and interferring
on delete or move operations when other syncs are stalled due to pending
move or rename operations.

These changes adapt the flag to be individualized per sync, whereas it
keeps the general flag to detect any possible problem with the new
approach.