Store timestamps debris folders below the current date parent folder
Every time a file or folder in the cloud is updated or deleted, the old
file is moved to the local debris folder.

When the filename is the same, in order to avoid name clashes, a new
folder with a timestamp and a counter (along with the date) is created
for each new file. This generates a lot of different folders for the
same day in the parent debris folder.

This commit adjusts the code so the folders with timestamps are nested
below the main folder with the date.

Before:
Rubbish/2024-04-09/MEGASync.log (version 1)
Rubbish/2024-04-09 19.32.27.0/MEGASync.log (version 2)
Rubbish/2024-04-09 19.39.10.0/MEGASync.log (version 3)

After:
Rubbish/2024-04-09/MEGASync.log (version 1)
Rubbish/2024-04-09/2024-04-09 19.32.27.0/MEGASync.log (version 2)
Rubbish/2024-04-09/2024-04-09 19.39.10.0/MEGASync.log (version 3)