SDK-3811. Fix RaidReq progress being greater than filesize
The real size of each raid part can be slightly smaller than the padded
part size (which is the size of the larger part size rounded to a
RAIDSECTOR - 16 bytes). When we count the progress, if the position for
a part is above the real part size, the difference must be substracted
to get the real data progress.

If the part is the parity part, it must take into account the real part
size of the unused part (which is the one being reconstructed with
parity).

This was already implemented. When the part pos was equal to the padded
part size, the difference was being subtracted. However, it wasn't being
taken into account the rare case where a part is still inflight, and the
part pos is greater than the part size, but still smaller than the
padded part size.

This commit fixes it and we know subtract the surplus.