Remove unnecessary log
This log is spammed when the soft limit is reached, as there is no flag
control to avoid it. It can happen when the throughput is, for example,
around 150KB/s, which is an acceptable speed and won't cause any issues.

We could include a flag to log it just once, but this log is not really
necessary as explained above: it will happen under expected scenarios,
when we change the soft time limit to the hard time limit, and adding a
new attribute to store a flag wouldn't really be justified.