Use recursive_mutex for mSyncVecMutex for safety
mSyncVecMutex was using a regular mutex. However, a lot of new locks
were added to different methods, including the ones for the main sync
thread, but there are others without this lock. There is a risk of
deadlocking and the mutex (mutual exclusion) protection must be
rethought. Meanwhile, this mutex is better to be recursive.