SDK-3855. Use PermissionHandler for testing directory permissions
- When folder permissions are removed and the test fails before they are
restored, it lead to errors when other processes (like Jenkins
maintenance) tries to remove those directories.

- This was fixed some time ago by creating a lambda function to check
the assert conditions and restore the permissions when the condition
failed.

- However, there were some additional problems:
1) This fix wasn't extended to the part of the test removing read
permissions. While it was the execution permissions the important ones
for directory removal, the lack of read permissions could also affect
cleaning processes if they need to list the directory contents before
removal.

2) Using ASSERT_TRUE within a lambda function doesn't finish the test
execution, because the ASSERT_TRUE causes a return from the function,
which is not the main test function (it's the lambda).

3) The using of chmod to remove and restore permissions, using flags for
the user permissions, caused the other permissions ('group' and
'others') to be removed. Notice that this always happens even when the
test doesn't fail. However, that's problematic

- In order to fix all these problems, a helper class PermissionHandler
has been defined to retrieve the original permissions, remove the
execute/read permissions without affecting the other permissions (group
and others), and restore the permissions upon destruction if they
haven't been restored at that point. This way, the test can finish due
to an ASSERT and the permissions will be correctly restored.