SDK-3807. Add sync errors to syncFolder method
MegaApi::syncFolder() returns a MegaError when the sync cannot be
synced. The MegaError object could also include a specific SyncError
field for sync errors. However, this SyncError wasn't being included to
the MegaError object returned to MegaApiImpl::addSyncByRequest()

These changes modifies the code returning the MegaError to include the
SyncError with it. It also adjust the documentation in MegaApi
accordingly, and improves the copy constructor of MegaErrorPrivate.

Note that the SyncError was actually being included to the MegaRequest
object (retrieved with MegaRequest::getNumDetails()), and this was
properly documented. However, it was confusing for developers, as there
is a specific MegaError being returned as a parameter within
MegaRequestListener::OnRequestFinish(), and it could be expected for
that MegaError object to include the SyncError (which is an optional
field of MegaError).