SDK-2620. Recreate sync tmp download folder upon API_EWRITE
When a new file is detected within a sync and it's going to be
downloaded, a tmp directory is created within the local debris folder.
This tmp folder is locked, and utilized for temporary transfer files.

This action is done only when the Sync::tmpfa (temporary FileAccess)
object is null, as it is defined during the tmp folder creation. If the
tmp folder is removed afterwards, the transfer will fail due to
API_EWRITE (the process cannot write on the target tmp folder as it does
not exist anymore). Even when the transfer is removed and the
sync-download is retried, the tmp folder will never be reconstructed
again, as the tmpfa object is still valid.

This commit fixes this problem by resetting the tmp transfer folder
(Sync::tmpfa) when a sync-download has failed several times (transfer
retried) due to API_EWRITE: after the transfer has failed it will be
reset and the sync-download will be retried from scratch. With the tmpfa
object being null, the tmp folder will be created again (if it doesn't
exist).