Refine SyncStallInfo::empty()
This method was checking whether syncStallInfoMaps (maps of SyncId -
StallInfoMaps) were empty.

However, there could be SyncIds/BackupIds with no stalls, because the
new code keep track of the noProgress counters to work properly (those
StallInfoMaps entries are purged later if they continue to be empty).

SyncStallInfo::empty() now calls StallInfoMaps::empty() for each entry
in the syncStallInfoMaps, which takes into consideration the cloud and
local maps - if they are empty, the entry is empty.