SDK-3206. Provide a blocked path when it is not possible to scan folder contents
When there is a file (or files) that cannot be accessed within a
directory, thus cannot be fingerprinted, this lead to the parent folder
to be scan-blocked: the scan state won't finish until this situation is
resolved.

Until now, there are at least two stall issues generated under this
scenario:
1) The stall issue for the file(s) that cannot be fingerprinted. Type:
FileIssue.PathProblem: CannotFingerprintFile.
2) The stall issue for the parent path which is scan-blocked. Type:
FileIssue. PathProblem: FilesystemErrorIdentifyingFolderContent

However, as these two stalls have the same type (FileIssue), and
contains the same dir path, they are simplified: the deeper path (the
file which cannot be fingerprintedd) is removed from the stall issues.

This lead to a situation where we only have one stall issue (FileIssue
with PathProblem:FilesystemErrorIdentifyingFolderContent), which is
generic and don't give information about the affected file(s), and we
don't have the important stall issue, which is the one giving
information about the affected file(s).

This commit fixes that situation by removing the scan-blocked stall
issue (PathProblem: FilesystemErrorIdentifyingFolderContent). It is
still generated when the folder itself is unreadable, but not when the
problem is because of its contents: we need the specific stall issues to
be aware about the problematic files and solve the problem.