Added getJSCDUserAttributes(...) family of functions.
The idea behind these functions is pretty straight forward: We want a
well-defined sequence of steps that causes the JSCD user attributes to
be created, a sequence of steps that we can easily compose with the
existing login process.

Now that we have these functions, we can introduce another, simpler
function, that will be responsible for injecting these attributes and
other necessary data into the sync engine as part of the login process.