Added injectSyncSensitiveData(...) and friends.
The idea of this function is simply to add an extra step before we let
the user know that their log in has succeeded. The reason we want to do
this is so that we can make sure that everything the sync engine needs,
exists and that it's been injected into the engine before anything crazy
starts happening.