Add SyncTest RemotesWithEscapedEscapesSynchronizeCorrectly
This new test adapts a big part of the disabled test
RemotesWithEscapesSynchronizeCorrectly.

The old test is kept as disabled because it tests a behavior that cannot
work ('f%30' and 'f0' cannot sync because they lead to a name conflict,
as both of them are 'f0'), however, part of the old test has been moved
to the new one.