Remove std::iscntrl(c) check within FileSystemAccess::unescapefsincompatible
When a path has been escaped with FileSystemAcess::escapefsincompatible,
control characters ('\7') have been taken into account for the escape.

When FileSystemAccess::unescapefsincompatible is later called, it needs
to be consistent with the previous call to escapefsincompatible. Right
now, it doesn't unescape control chars, and the result is not consistent
between calls.

The reason to avoid unescaping control chars was to prevent creating
such characters in the cloud, but it can be done anyway (there are no
restrictions to create those files in the cloud), and it generates
problem for filesystems whose LocalPaths are affected by
FileSystemAccess::escapefsincompatible, due to the inconsistency.