SDK-4044. User intervention stall issues must never be triggered for backups
- Some cases has been reported where an user intervention stall issue
was triggered for a backup. Specifically,
LocalAndRemotePreviouslyUnsyncedDiffer

- Several checks has been added to avoid calling
resolve_userIntervention for backups

- Added logs and asserts for different scenarios

- In the rare case where we find again a CXF (cloud node and FS node
exists, they are different, and there isn't a sync node) a new sync node
will be created from the FS node, later resulting in an upsync to fix
the mismatch.

- In case of CSF (cloud node and sync node and FS node exists) with all
of the nodes different between them (this should never happen for a
backup), or with only the cloud node different (this shouldn't happen
either, because the cloud cannot be changed for a backup), it will be
resolved with an upsync.

- Added a log to properly track when the mirroring state is set for
backups (mirroring mode is used for the first time, or if the backup was
disabled due to errors or if it is an external backups), and the
transition to monitoring state.