Working version of getAllNodeTags
This commit implements a not very efficient method relying on the
`search()` utility to get all the nodes and then processing their tags
to return the list with the unique ones.

A new test case is also introduce to be used as a refactoring
validator.

At this point, the search string is supposed to match exactly the nodes
tag (as it is implemented for the `MegaSearchFilter::byTag` option)