Added missing include in sqlite.h
This header file starts with an `#ifdef USE_SQLITE` statement. The
language server doesn't know if this is defined or not if we don't
include the `mega/config.h` file. The safest way to include it is
through `mega/types.h`.

This was compiling fine because of when this file was included, it was
done in places where the `config.h` was included previously.