Eagerly prune exited sessions from the MountDB.
The reason for this change is that when the system is under heavy load,
the application may receive a ton of queued FUSE requests from the
kernel when a given FUSE connection is being torn down. This is
problematic because the existing code, when receiving a request, checks
whether the connection has exited and if so, queues the destruction of
any associated mount.

The result of this is a giant storm of "destroy" tasks which, depending
on thread availability, may not be able to be processed which can result
in a deadlock.

This change resolves the issue by ensuring that sessions are removed
from the MountDB's "request loop" as soon it's been determined that they
have exited: Associated mounts are still destroyed but such a "destroy"
task is only emitted once.