Uploads no longer begin as soon as they are created.
Instead, we expect whoever creates an upload to explicitly begin that
upload. The reason for this change is that in some bizarre cases, it's
possible that the thread servicing the MegaClient can start and complete
the upload before ClientAdapter::upload(..._) returns to its caller.

This is problematic as it was expected that ClientAdapter::upload(...)
would always return before the upload it started actually completed.
That is, the callback the user provides is expected to be able to make
use of the UploadPtr that ClientAdapter::upload(...) returns.