Allow combining with or text conditions in the query
To implement this feature, a whole refactor of how the "WHERE" parts of
the queries handles the filter was performed. Now, a single function is
called to decide if a node from the query is valid or not. In this way,
the conditions can be handled programatically in the c++ side leading in
a more maintainable code.

Also, this new filter function was reused both in the search and
getChlidren functionality, improving even more the maintainability.

During the refactor, a lot of utilities were developed that might be
useful in other place of the code.