Get rid of the NodeSearchFilter::copyFrom<T> method
This was a template method but it was only used with MegaSearhcFilter
objects so a new static free function was introduced in megaapi_impl.cpp
to perform this transition from outer to intern layer. For that, some
helper setters needed to be defined in NodeSearchFilter.

Also some helper comments where added in sqlite.cpp