Give the client some time to import the Welcome PDF.
This isn't a super robust way to check but it ensures that the client
receives APs for an import before it logs out.

If we're operating in PWM or VPN mode and the APs are not being
filtered, the test will assert.

That is, we want to be sure that APs *are* being filtered.