Append to log files rather than overwrite them.
This is necessary in some circumstances where the logger has been torn
down but some components of the system continue to log messages.
Ideally, we could ensure that all components of the system were truly
terminated before killing the logger but sometimes that's impossible or
difficult to do.

One example of this would be when we use GTest's fixtures and test
suites with static members.