getTagPosition(...) no longer escapes its pattern.
Callers are expected to escape their inputs if they don't want the
search to process wildcards.

The reason for this is simple: Most of the time, we don't want to
interpret wilcards but sometimes, we do. An exampple of this would be
when we want to perform partial matching of tags.