Avoid gerenating FileChangingFrequently stall issue if FileChangingState is not initialized
In case movement/renaming is detected for a LocalNode, we need to check
if there are still pending changes in flight or that LocalNode changes
so many frequent.

In case FileChangingState is not initialized yet checkIfFileIsChanging
will return true so a FileChangingFrequently stall issue will be
generated.

This fix avoid to generate the stall issue in case we detect that
FileChangingState is not initialized.